# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs for developers",
        "contact": {
            "name": "Leonardo Aragão",
            "url": "https://github.com/lewoaragao",
            "email": "lewoaragao@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registering a new User.",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of User.",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "The email of User.",
                                        "type": "string",
                                        "example": "john@doe.com"
                                    },
                                    "password": {
                                        "description": "The password of User.",
                                        "type": "string",
                                        "example": "@api4devs"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Starts User session.",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "The email of User.",
                                        "type": "string",
                                        "example": "john@doe.com"
                                    },
                                    "password": {
                                        "description": "The password of User.",
                                        "type": "string",
                                        "example": "@api4devs"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "access_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log out of the User session.",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                }
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Listing all resources",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "Response success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful action!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Storing a new resource.",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of Product",
                                        "type": "string",
                                        "example": "PC"
                                    },
                                    "description": {
                                        "description": "Description of Product",
                                        "type": "string",
                                        "example": "Personal Computer"
                                    },
                                    "price": {
                                        "description": "Price of Product",
                                        "type": "decimal",
                                        "example": "200.50"
                                    },
                                    "quantity": {
                                        "description": "Quantity of Product",
                                        "type": "integer",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show a specific resource.",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identification of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful action!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Updating a specific resource.",
                "operationId": "d3467c795490afbb49680821c9bd6b0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identification of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of Product",
                                        "type": "string",
                                        "example": "PC"
                                    },
                                    "description": {
                                        "description": "Description of Product",
                                        "type": "string",
                                        "example": "Personal Computer"
                                    },
                                    "price": {
                                        "description": "Price of Product",
                                        "type": "decimal",
                                        "example": "200.50"
                                    },
                                    "quantity": {
                                        "description": "Quantity of Product",
                                        "type": "integer",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful action!"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Deleting a specific resource",
                "operationId": "c830155516bad1f72b0081160edeb2ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identification of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successful action!"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Response Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "description": "Schema of Model Product",
                "properties": {
                    "id": {
                        "description": "Identification of Product",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of Product",
                        "type": "string",
                        "example": "PC"
                    },
                    "description": {
                        "description": "Description of Product",
                        "type": "string",
                        "example": "Personal Computer"
                    },
                    "price": {
                        "description": "Price of Product",
                        "type": "decimal",
                        "example": "200.50"
                    },
                    "quantity": {
                        "description": "Quantity of Product",
                        "type": "string",
                        "example": "100"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "Schema of Model User",
                "properties": {
                    "id": {
                        "description": "Identification of User",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of User",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email of User",
                        "type": "string",
                        "example": "john@doe.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}